feat(FEAT-031): complete phase 1 - Vitest TS test relocation (FR-2)
