+++ tests/unit/capability-discovery.test.ts
