+++ tests/unit/git-utils.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest';
-import { filterNoiseCommits, type ParsedCommit } from '../lib/git-utils.js';
+import { filterNoiseCommits, type ParsedCommit } from '../../scripts/lib/git-utils.js';
 
 function commit(raw: string, overrides?: Partial<ParsedCommit>): ParsedCommit {
   return {
