+++ tests/unit/qa-feat-030-contract.test.ts
