+++ tests/unit/shared-scripts.test.ts
